prep for auto-typing DefMacro arguments
