+++ rtx_package/src/package/mod.rs
@@ -8,11 +8,11 @@ pub use std::collections::VecDeque;
 pub use std::sync::{Arc, RwLock};
 pub use log;
 
+pub use rtx_core::*;
 pub use rtx_core::common::dimension::Dimension;
 pub use rtx_core::common::mudimension::MuDimension;
 
 pub use rtx_core::aux_macros::*;
-pub use rtx_core::common::error::*;
 pub use rtx_core::common::float::Float;
 pub use rtx_core::common::font;
 pub use rtx_core::common::font::Font;
