+++ rtx_package/src/package/setup_binding_language.rs
@@ -1587,6 +1587,52 @@ macro_rules! DefMacro {
   }};
 
 }
+
+#[macro_export]
+macro_rules! TypedMacro {
+  // closure
+  ( $cs:ident $($var:ident : $ptype:ident),+ => sub [ $gullet:ident, $inner_state:ident ] $body:block $($input:tt)*) => {
+    let cs_token = concat!("\\", stringify!($cs));
+    TypedMacro!(T_CS(cs_token) $($var : $ptype),+ => sub [ $gullet, $inner_state ] $body $($input)*)
+  };
+  ($cs:literal $($var:ident : $ptype:ident),+ => sub [ $gullet:ident, $inner_state:ident ] $body:block $($input:tt)*) => {
+    TypedMacro!(T_CS($cs) $($var : $ptype),+ => sub [ $gullet, $inner_state ] $body $($input)*)
+  };
+  ( T_CS($cs:expr) $($var:ident : $ptype:ident),+ => sub [ $gullet:ident, $inner_state:ident ] $body:block $($input:tt)*) => {{
+    let options = defi_opts!(@munch ($($input)*) -> {ExpandableOptions,});
+    let mut parameters = Vec::new();
+    $(
+    parameters.push(
+        Parameter {
+          name: stringify!($ptype).to_string(),
+          ..Parameter::default()
+        }
+        .init(outer_state!())?,
+      );
+    )+
+    let expansion_closure: Option<ExpansionBody> = Some(ExpansionBody::Closure(Arc::new(
+      move |$gullet: &mut Gullet, mut args: Vec<ArgWrap>, $inner_state:&mut State| {
+        $(
+          let $var: $ptype = match args.remove(0).try_into() {
+            Ok(v) => v,
+            Err(e) => {
+              Error!("expected", "argument", $gullet, None, e);
+              $ptype::default()
+            }
+          };
+        )+
+        WithInnerState!($body, $inner_state).into_tokens_result()
+      }
+    )));
+    let params = if parameters.is_empty() {
+      None
+    } else {
+      Some(Parameters::new(parameters))
+    };
+    defi_macro!(T_CS!($cs), params, expansion_closure, Some(options));
+  }};
+}
+
 /// Internal auxiliary, only purpose is to bind state, then call the api::def_macro
 /// function, where the interior macro installation logic resides.
 #[macro_export]
